requirements.txt
subprocess
requests
pytube
langchain_community
langchain
youtube-transcript-api
python-dotenv